pyproject.toml
[project]
name = "qbench"
version = "0.1.0"
description = "QBench - Queue Management Benchmark for AI Agents"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0.0",
    # A2A and AgentBeats dependencies (optional for research use, required for competition)
    "a2a-sdk>=0.3.5",
    "fastapi>=0.115.0",
    "starlette>=0.48.0",
    "sse-starlette>=2.0.0",
    "uvicorn>=0.35.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    # Purple agent dependencies
    "litellm>=1.0.0",
    "loguru>=0.7.0",
    # TOML support for Python 3.10
    "tomli>=2.0.0; python_version < '3.11'",
]

[project.scripts]
qbench = "qbench.cli:main"
qbench-run = "agentbeats.run_scenario:main"
qbench-eval = "qbench.evaluator.qbench_evaluator:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/qbench", "src/agentbeats", "examples"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # imported but unused
"tests/**/*.py" = ["B011"]  # do not use assert False

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v"
